Implement Authorization using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -69,7 +69,17 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        // if(\Gate::allows('update-question', $question)) {
+        //    return view("questions.edit", compact('question'));
+        // }
+        // abort(403, "Access Denied");
+
+        if(\Gate::denies('update-question', $question)) {
+           abort(403, "Access Denied");
+        }
+
         return view("questions.edit", compact('question'));
+
     }
 
     /**
@@ -81,6 +91,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)) {
+           abort(403, "Access Denied");
+        }
+
         $question->update($request->only('title', 'body'));
         return redirect('/questions')->with('success', "Your question has been updated");
     }
@@ -93,6 +107,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('update-question', $question)) {
+           abort(403, "Access Denied");
+        }
+
         $question->delete();
         return redirect('/questions')->with('success', "Your Question has been deleted.");
